GoogleCloudStorage Live implementation working w/ Arc_gcs
We are using `store/2` directly, instead of using Arc as an attachment
to a "model". As such, we only define a _private_ GoogleCloud module
inside the Live module so it can `use Arc.Definition`.

Had to use the github version of `:arc`, since the one on Hex.pm didn't
allow to specify a different folder via configuration.

Added the `secrets` folder to .gitignore as well.

GoogleCloudStorage configuration lives on the ElixirDrip umbrella
application.