Download pipeline definition with GenStage
The Starter module, as the Producer of the pipeline, is using a polling
mechanism to check if the queue it uses has more events to handle the
demand it received.

We will later remove the polling mechanism in favour of the QueueWorker
simply telling the Starter producer that events were enqueued.