Instead of a prepare function "ad-hoc", we use a proper callback
This callback has a default implementation. Since we annotated the
`init/1` as @impl, because it's in fact the implementation of the
`GenState.init/1` callback, we needed to add the `@impl` annotations to
every overridden callback.

This has the extra benefit of the compiler warning us if there's no
callback to override in the first place.