Add Accounts context with User schema
For now, the User schema just has the username and password_hashed
fields. More will be added according to what's needed. Also added the
`comeonin` library as a dependency, so that we save the passwords of the
users hashed, and can later verify them (through the
`verify_user_password/1` function).