Many to many Media<->Owner association
New `MediaOwners` and `Owner` schemas that point to the `media_owners`
join table and the `users` table, respectively.

This way we can use the `put_assoc/3` function when storing a new user's
file, to store the file associated with a given user.

Also refactor of the changeset functions to the `Media` module,
instead of living under the `Storage` module.