Storage.allows `retrieve` and `share`
Retrieve checks if the User has access to the media, before downloading
it.

Share adds an entry to the `MediaOwners` schema, by checking the unique
constraint first.

Storage has a partial `user_media_query` that joins the `Media` and
`MediaOwners` schemas.

Given an explicit `:single_share_index` to the `unique_index` we have
created on the `media_owners` table.