Running DB migrations with a MIX_ENV=migration
Since the release doesn't include Mix, we can't rely on it to run the DB
migrations with the release. To overcome this, we need to create a
custom module that will run the migrations for us and will be included
in the release.

Since our goal now is to simply deploy a working ElixirDrip version, we
are still running the migrations with the full code (that still lives in
the deployed container).

Afterwards, we can create a two-step build that will only deploy a
container with the release as it should be, thus it will need to run
migrations via the mentioned custom module.